Implementation of hmmscan variant for DNA search, currently in a separate executable called nhmmscan. This may at some point in the future be merged into hmmscan. Implementation required a little reshuffling of functions, specifically removing the requirement of having a P7_HMM passed in to reparameterize_model(), which is the cause of many of the minor file changes.
The new nhmmscan scales pretty well up to 8 cores (as much as I've
tested so far). Haven't done any optimization to see if this can
be improved.  Below, see the Dfam hmm library vs the example  
sequence we currently use on the Dfam website (~14Kb) :

cpu    user      waste
0   0m55.989s 
1   0m49.853s      -
2   0m24.994s      0%
4   0m12.862s    3.2%
6    0m9.242s   10.0%
8    0m7.189s   13.3%

While doing this, and comparing output to nhmmer, I noticed a flaw 
in my reparameterization scheme (which I implemented to limit the   
tendancy of composition bias to cause serious over-extension issues).
The method previously smoothed the local window composition with the
composition from the large sequence block (default 1/4Mb) from which
the window was extracted. This had the unfortunate byproduct of
allowing a change (e.g. addition or removal) in sequence in one part
of a file to impact scores of hits 10s of Kb away. Reverted to
something that had worked essentially as well in earlier tests,
smoothing with the default background frequencies. The result is
agreement between nhmmer and hmmscan bit scores for comparable
searches, as well as consistency between similar nhmmer runs.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4212 db3e70e4-12e1-0310-b23f-81a6e3d09ba7