libhmmer.a now compiles cleanly.
The pipeline is (temporarily, we hope) using a disgusting workaround
when in hmmscan mode.  See workaround_get_profile() and documentation
around it, for plan to refactor it when 3.1's dust settles a bit.  The
issue is that we need a standard P7_PROFILE in our new pipeline, but
hmmscan reads only vectorized P7_OPROFILE pieces from its hmmpress'ed
datafiles; the workaround reads an entire HMM from the .h3m file, and
converts it to a P7_PROFILE, which is inefficient both in alloc/free
of the HMM and in HMM->profile conversion time.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4279 db3e70e4-12e1-0310-b23f-81a6e3d09ba7