Manually applying bug fix from trunk revision 4281 to hmmer's Infernal 1.1 release branch. The fix is simple, involving a single line change of a flag value in make_post_msa(), to make hmmbuild -O work properly when used in combination --hand and when nongap RF columns include 100% gaps in the aligned sequences.
This bug was minor and would have been very rarely encountered in the
Infernal 1.1 version of hmmbuild. I mainly wanted the fix in the 1.1
branch because the hmmer trunk is still unstable and I want to use
'hmmbuild -O --hand' to generate alignments with truncation annotation
to test changes I'm making to esl-alimerge.  


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/infernal@4285 db3e70e4-12e1-0310-b23f-81a6e3d09ba7