'daemon' subdir now compiles with examples and benchmark.
Build system now allows a directory to contribute some objects to
libhmmer.a (listed in ${LIBOBJS}), or keep some objects in the local
subdir (listed in ${OBJS}), or both. This allows a subdir to consist
of a standalone, libhmmer.a-dependent set of objects and executables:
the daemon, for example. The daemon's implementation does not put any
symbols into the rest of HMMER3.

To make this work, because of a bad interaction between 'make' and our
technique for including drivers (such as unit tests) in the same .c
file they're for, we also have to create a lib${MODULE}.a static
library in each subdir that has non-empty ${OBJS} list. The problem is
that you can't 'cc -DFOO_TESTDRIVE foo.c ${OBJS}' when ${OBJS}
contains foo.o: the linker will barf on duplicate symbols (foo.c and
foo.o). But you can 'cc -DFOO_TESTDRIVE foo.c -lbaz' where libbaz.a
contains foo.o, because the library linker is smart enough.  So, our
src/*/Makefile.in's define ${LIBTMP}, ${LIBTMPDIR}, and ${LIBTMPDEP} when
they have a nonempty ${OBJS}. See the daemon/Makefile.in for an example.

What we seem to be headed for is different kinds of src/ subdirs:
subdirs consisting only of libhmmer.a code (and tests, benchmarks,
etc), and subdirs consisting of executables that depend on libhmmer.a,
including more complex executables that have several .c/.o files of
their own.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4293 db3e70e4-12e1-0310-b23f-81a6e3d09ba7