Bugfix in makenhmmerdb.c.  r4010 changed log2() calls in makenhmmerdb.c to log() * eslCONST_LOG2R, for portability -- log2() is not portable. However, this change caused the i020-fmindex-core.pl test to fail consistently on all platforms. I initially thought it was a more subtle and FreeBSD-specific failure.
Problem is an integer roundoff issue. makenhmmerdb was taking log2()
of an integer power-of-two and storing it in an integer, but log2() is
double log2(double x), so we required a floating point conversion of
the arg, and an integer conversion of the result. This is way too
trusting of floating point arithmetic. Though a sensible log2()
implementation might well return an integral answer for a
power-of-two, I doubt it's required to (as opposed to being off by an
epsilon); in any case, it certainly isn't guaranteed to work when we
use log(). i.e. log2(8.0) ought to be 3.0, but it might be 3.0-epsilon
which truncates to 2. I jammed (int) round() calls in there to make it
work, but it really should be refactored cleanly.

(inspection shows makenhmmer has other issues which will someday bite
us. For example, it's doubly setting defaults, both in its options
structure and in main().)

fixed a typo in i020-fmindex-core.pl; it was deleting <tmp>.fm twice,
and not deleting <tmp>.fa.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4311 db3e70e4-12e1-0310-b23f-81a6e3d09ba7