Modified nhmmer so that the database range-restriction flags are now    --restrictdb_stkey    --restrictdb_n and are "secret" flags (visible in the source, but not shown with -h).
Previously the beginning of the range was specified by an int, n, to 
request starting at the sequence with ordinality <n> in the target. 
Now, it accepts the sequence name (key) instead. That's because the
ssi primary key is sorted lexicographically, and original order is
lost. A data structure change could get us there, but for now, using 
the key is sufficient for our purposes. This is not enabled for MPI.

Also implemented the flags in hmmsearch and phmmer. (I can push the
change to jackhmmer, too ... will discuss with Rob and Sean).



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/nhmmer@4401 db3e70e4-12e1-0310-b23f-81a6e3d09ba7