Important bugfix in 3.1 branch, for problem reported by Jody.
scores_per_pos element in P7_DOMAIN structure, new to 3.1, was not
being allocated, initialized, reallocated, free'd properly. Fixes in
p7_domaindef.c and p7_tophits.c. Most important one (and most directly
responsible for the bug Jody saw) is in p7_tophits_Destroy(), where
code was free'ing dcl->scores_per_pos, instead of
dcl[j].scores_per_pos, for j < ndom.  This results in a memory leak,
and on pathological rare sequences where ndom==0, a spurious free if
scores_per_pos happens to be non-NULL... which only happens as a
stochastic, race-like condition in multithreaded code, probably
because of the other issues with initializing it properly.

This is not a problem in H4 trunk, where P7_DOMAIN routines were
already refactored.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/3.1@4413 db3e70e4-12e1-0310-b23f-81a6e3d09ba7