The three hmmd files (mstr, wrkr, and c2) contain the line: #include <arpa/inet.h>
On my system, and apparently all the others we typically test, 
/usr/include/arpa/inet.h contains:
   #include <netinet/in.h>
which defines the struct 'sockaddr_in' used in the hmmd files.

This does not appear to be the case for OpenBSD - the inet.h 
file doesn't include the in.h file. The fix is to directly 
include netinet/in.h within those three files. 

(The missing include was pointed out by 
senthil.murugapiran@unlv.edu)



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/3.1@4482 db3e70e4-12e1-0310-b23f-81a6e3d09ba7