Many small variable name changes (from msv to ssv), and some architectural changes to get FM closer to nhmmer incorporation.
One major change is a flattening of the file hierarchy. I was
having a great deal of trouble with circularity errors in
compliation, as the FM code depended on hmmer.h, and the
hmmer code depended on fm.h.  This can eventually be smoothed
out by pulling true FM-only code out into an FM library, 
but for now, just move all fm_x files to the main directory.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/3.1@4511 db3e70e4-12e1-0310-b23f-81a6e3d09ba7