Bug #h101 fix. hmmsearch segfaults if DNA HMM is searched against a protein db.
Problem is that <dbfp> is not getting set into digital mode, either in
serial_master or in mpi_worker in hmmsearch.c. In serial/threaded
mode, code was erroneously reaching into dbfp to set dbfp->abc, but
needs to call esl_sqfile_SetDigital(), which not only sets dbfp->abc,
but also more, including (crucially) setting the legal <inmap>. In MPI
mode, wasn't even doing that.

(This is the patch to the 4.0 trunk.)



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4516 db3e70e4-12e1-0310-b23f-81a6e3d09ba7