Fixed edge case in which makehmmerdb was overwriting the counts in the last of block count arrays, which eventually caused crashes in some cases. The underlying issue was funky behavior by the built in ceil() function. It takes a double, and I mistakenly was passing a float ... and ceil() rounded down instead of up. Resolved.
Fixed a couple other edge cases in the BWT interval calculation.

nhmmer with FM now successfully runs rmark3, and gives very good 
sensitivity. The couple decent hits found by the standard nhmmer
and not FM nhmmer are caused by hits having perfect pathological
conditions to break the heuristic, not bugs.  Speed is on the
order of 2.5x faster so far, and it seems to scale to at least
8 threads.   



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/3.1@4538 db3e70e4-12e1-0310-b23f-81a6e3d09ba7