Minor code cleanup.
Fixing a bug that cropped up in 5 out of 1/4 million searches of the human
genome with CNE models from Gill Bejerano. The cause was that the alignments
had very high gap frequency and lengths, leading to extremely long MAXL
(e.g. 70K+ for a model with 24 match states), which exposed an edge case bug 
in the way nhmmer splits up very long windows in order to avoid numerical 
stability issues in Fwd/Bkwd. I've (a) fixed that bug, and (b) modified 
the MAXL calculation to limit MAXL to no more than 20*M, where M is the # 
of match states. 



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/3.1@4564 db3e70e4-12e1-0310-b23f-81a6e3d09ba7