Fixed a rare unit test failure in sparse_fwdback_utest. Was reproducible with   ./sparse_fwdback_utest -s 1705952920
The "compare reference" and "reference constrained" unit test were
being too stringent by requiring that the reference and sparse DP
Viterbi trace are exactly identical. 

Although one might think that it would be nearly impossible for a
profile/sequence alignment to have two or more identically-scoring
alternatives for a Viterbi trace, it's only mostly impossible. The
case that was arising sporadically in nightlies is when the Viterbi
path has a single residue in a local alignment to a single match
state. Now all occurrences of the same residue in the target sequence
give an identically-scoring Viterbi path. For example, if ML_42
aligned to Y65 is the best possible alignment, then any other Y in the
sequence gives exactly the same score, via SN...NB->ML_42->EC..CT
trace, because t_N->N == t_C->C by construction and ML_42 gives the
same emission score for Y regardless of where the Y is in the target
seq.

The fix is to implement a less stringent version of
p7_trace_Compare(), here called trace_compare_loosely(), which
verifies that the two traces have exactly the same sequence of M/I
emitting states, and that they emit exactly the same subsequence of
residues (by identity, rather than by idx in target sequence).




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4568 db3e70e4-12e1-0310-b23f-81a6e3d09ba7