exp-decoding experiment driver written. This will enable a set of experiments comparing ensemble-based calculation (such as most probable labeling, MPL) to alignment-based calculation for the purpose of identifying domain locations.
The exp-driver code includes acceleration_filter(), a stripped down
and simplified version of the p7_Pipeline(), implementing only the
vector acceleration "front end" of HMMER, as I've often described it
in slides as a separable black box in front of "domain
postprocessing", the main focus of H4 development. The
acceleration_filter() code might be useful for pedagogy, or even as a
basis for simplifying the code in p7_Pipeline() someday.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4590 db3e70e4-12e1-0310-b23f-81a6e3d09ba7