Bug #h104 fix. Qing-dong Zeng provides an example H2 save file that H3 won't read. H3 claims bad format. The save file appears to be from a 1998 version of SMART.
Two problems identified. One is that the file has a DESC
line but no description. Fixed p7_hmmfile.c::read_asc20hmm()
to allow this.

Second is that hmmemit (only?) complains "whoops, HMM is bad!",
because hmmemit does a p7_hmm_Validate() on the model, and it's
slightly nonnormalized. Fixed by adding a p7_hmm_Renormalize() call to
the end of read_asc20hmm().



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4597 db3e70e4-12e1-0310-b23f-81a6e3d09ba7