Making a commit for Jody, who seems not to have commit privileges:
Adding change to return taxonomy ids from hmmpgmd.
    
Adding an additional string to the sequence headers in hmmpgmd.fa
produces something like the following:
    
>1 000101001 12343829483298 1234
    
This change to the hmmdwrkr allows us to take the last numerical value
(1234) and place it in the acc field of the returned binary structure.
The current use for this is to place the taxonomy id in the results, so
that we don't have to do a tree lookup for each of the sequences.
    
Note: Since the strtol function is being used to extract the value from the
header, it can only return numerical values, should one wish to use this
for some other means.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/3.1@4628 db3e70e4-12e1-0310-b23f-81a6e3d09ba7