Identified a design flaw that affects posterior decoding of glocal alignment paths: the 'mute partial cycle flaw'.
Briefly, the flaw is that it is possible to visit the same DGk cell
twice on the same row i of a decoding matrix, when a DGk...E exit
suffix of one domain overlaps the G->DG1..DGk->MGk entry prefix of the
next domain. When Decoding does wing unfolding, it overcounts the
overlapped DGks. Forward/Backward are not subject to the flaw, because
wing retraction (by design) avoids mute cycles, even these partial
ones. The flaw appears to be negligible, both on analytical grounds
(SRE:J13/60) and based on the continued success of unit tests that are
capable of detecting it. It's negligible because there's always a
better single domain path that uses the overlapped DGks just once on a
shorter all-internal deletion, rather than the two domain path that
exits on a long DGk suffix and re-enters on a long DGk prefix.

However, it may become an issue in the future, so we need to remember
it.

Added the utest_mute_partial_cycle() 'test', which creates a specially
crafted profile/seq comparison that demonstrates the flaw, and
produces a decoding matrix that fails validation.

Added validate_probs() to p7_refmx_Validate() for decoding matrix, to
assure values v are 0 <= v <= 1+eps. (+eps, because we have fp error
accumulation in play.)

Added utest_randomseq() to test suite for reference_decoding. This
test is capable of detecting the mute partial cycle flaw, especially
for small L,M, but it does not; that's evidence that the flaw is
negligible.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4656 db3e70e4-12e1-0310-b23f-81a6e3d09ba7