Unit tests for ASC Decoding are in place, and they pass.
There are 7 tests. One is the "overwrite" test, as in
reference_decoding. The other 6 correspond to the fwdback tests,
extended to posterior decoding, with various contrived
profile/sequence comparisons that force a standard calculation to use
an anchor set, so the ASC calculations can be exactly compared to
standard DP.

I believe reference ASC Decoding is now working properly.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4657 db3e70e4-12e1-0310-b23f-81a6e3d09ba7