Added --seed_consens_match flag to nhmmer. A run of <n> consecutive consensus-matching letters will override the score threshold requirement, and automatically require SSV extension. Default <n> is currently 10, though I may shift it to 11 later.
This was done after testing with single sequence queries highlighted the fact that
their low scores could cause consensus-matching blocks to be removed due to
low score density. This adds marginal overhead, with probably a good return 
in sensitivity (to be tested).

Also fixed an error in which nhmmer was building HMMs for input alignments
using a seed of 0, regardless of the default (42) or an override via
flag.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/3.1@4658 db3e70e4-12e1-0310-b23f-81a6e3d09ba7