Improved the MEG alignment implementation. It's now filling in the entire matrix, even where paths aren't valid -- it's the reference implementation, so we should care more about clarity, not efficiency. Invalid cells are -inf, which clearly distinguishes them from cells that are valid but unreached, which are 0.0.
Also, it's now dealing with local/glocal in the fill stage. Glocal
envelopes use G/MG/DG/IG states, local envelopes use L/ML/DL/IL. Same
rationale: though we could save 2x in memory, the reference
implementation puts a premium on clarity. 

(That said, the DP algorithms are complicated, and made more so by the
pointer-based access patterns in both the profile and the matrix.)

Renamed to "anchor and envelope-constrained" (AEC) alignment. If
there's ever a true ASC alignment algorithm, will want to distinguish
this one-matrix version, which is constrained by envelope bounds ia/ib
as well as by anchor i0. 

"AEC" refers to the constraint on valid cells in the DP matrix, and on
the valid path ensemble. "MEG" refers to the optimization criterion.
So this is AEC+MEG. We could imagine other combos, such as
AEC+Viterbi, etc.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4663 db3e70e4-12e1-0310-b23f-81a6e3d09ba7