In P7_SPARSEMASK, I decided it would be advantageous to flank the segment coords with sentinels, and also to make them a little more clearly an ia/ib pair for each segment. Segments are now numbered 1..S, rather than 0..S-1, with sentinel values at 0,S+1. This will simplify the structure of some ASC sparse code. However, it's a dangerous global change, risking introducing off-by-one problems in code that I've already checked.
sm->i[] (which was an unstructured list of ia,ib pairs) is now an
array of structs, sm->seg[1..S].ia,ib; nseg is now S; ialloc is now
salloc; n_irealloc is now n_srealloc. Sentinels for seg[0] are -1; for
seg[S+1], L+2.

Passes unit testing, and I think it's ok, but I'm checking this in as
a discrete diff, just in case I need to revert it.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4670 db3e70e4-12e1-0310-b23f-81a6e3d09ba7