p7_sparse_AnchorsSeg(): Segmental divide and conquer version of the MPAS algorithm is complete and appears to work.
A landmark, because this is pretty much the most difficult
implementation I need in the remaining HMMER4 road map, and for a
while it looked like it might be impossibly difficult. Sparse ASC
Backward implementation remains to be written, but that will only be
ridiculously difficult. Then it should be all downhill from there.

"Appears to work" means it works on several anecdotal examples, giving
the same answers as the sparse MPAS complete-seq version. Still needs
unit tests. And speed benchmarking - the immediate critical question
is whether segmental D&C MPAS is fast enough. If not, I'm out of
fallback positions, this is the final stand for the last year or so of
work. MPAS is almost certainly an NP-hard problem and it could well
prove to be infeasible in the end. But I don't think so. I think I've
got it now.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4720 db3e70e4-12e1-0310-b23f-81a6e3d09ba7