HMMER converted to BSD licensing, from GPLv3.  COPYRIGHT file merged into LICENSE, to simplify.
Have been meaning to do this for a long time. I've searched the code
and I don't believe we have any GPL-licensed code from outside
contributors. All contributions are compatible with BSDL, so we are
free to re-license.

If anyone sees otherwise, let me know!

The only complication I can think of is that our sublicense to Farrar
patent-pending SIMD technology was granted on the understanding that
redistribution of the technology would be limited by the GPLv3 patent
clause. However, Michael's patent never issued.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4735 db3e70e4-12e1-0310-b23f-81a6e3d09ba7