Bug #h108 fix: nhmmer corrupted by * residues in input seqs.
Original report was for nhmmscan dying with fatal NaN exception in
fwdback.c. Narrowed to a test case of a DNA sequence containing a '*'
nonresidue character. Three problems fixed; 2 are in an Easel commit,
and the HMMER change is:
  
p7_oprofile.c::p7_oprofile_UpdateFwdEmissionScores() - this seems to
have been the actual problem. It was recalculating scores and
passing vectors to esl_abc_FExpectScVec() for dealing with
noncanonical residues. esl_abc_FExpectScVec() documentation says
that K, Kp-1, Kp-2 must be set by caller; UpdateFwdEmissionScores()
was not doing that.  Result: uninitialized scores for -, *, ~.
  


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/3.1@4782 db3e70e4-12e1-0310-b23f-81a6e3d09ba7