Removing a trap in nhmmscan and nhmmer that disallowed aligned sequence files as target inputs. When I originally put this in place, I was avoiding an infinite loop in esl_sqio_ReadBlock() when reading aligned files.
I tracked the error to sqascii_ReadWindow(), in the section for aligned 
sequence. The code simply uses esl_sq_FetchFromMSA() to sequentially 
grab sequences, which get appended to the active block in the calling 
function. I was reusing a bit of code that checked for the possibility
that only part of the sequence had been read (possible when reading
windows of long sequences, but not possible with MSAs). The method 
of reuse allowed the sequence id to remain unincremented, so that
sqascii_ReadWindow() kept returning the first sequence of the MSA
(causing that sequence to be searched over-and-over again) until the 
program was killed.

With this infinite loop fixed, I've now removed the trap.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/3.1@4784 db3e70e4-12e1-0310-b23f-81a6e3d09ba7