Adding two new fields to the P7_DOMAIN struct,  iorf and jorf.
These will be used in Walt's translated search routine to capture the range 
in the DNA sequence of the ORF containing the match to a protein query. I'm
adding them here so that I can give the HMMER web team a preview of the 
changes to the binary that are required for the translated search update.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/3.1@4800 db3e70e4-12e1-0310-b23f-81a6e3d09ba7