Use correct sequence name, accession and description instead of those from ORF
Copy the name, accession and description from the nucleotide sequence into
the target sequence. This target sequence is an ORF orginally returned from
do_sq_by_sequences in gencode.c which is passed into the pipeline for processing,
and contains the ORF name, accession, and description which we do not want to use.