Adopting git flow: first commit on h3-master branch.
Updated version to 3.1b3 for the moment, though this is a placeholder
on our new master branch. Updated copyright info. Moved README to
README.md and reformatted so it will look nice on github. Changed from
GPL to BSD license.

I believe there is only one thing to note with moving from GPL to BSD,
and that's the Farrar SIMD patent license. We were granted a
nonexclusive license provided that the code is distributed under
GPLv3, with its patent clause. Farrar died in 2010 and the patent
never issued, making it a moot point.