Fixes bug in nhmmscan, which segfaults on old model files.
nhmmscan depends on reading MAXL field from profile save file, but
pre-3.1 save files don't have it, and nhmmscan didn't check -- and in
the SSV_Longtarget() was validating the value of om->max_length before
using it to set om->tjb_b, underflowing, and creating havoc
thenceforth.

Added a check in nhmmscan that validates that at least the first model
file has max_length field. This should suffice for now but it is not a
bulletproof solution -- I believe it's possible to have concatenated
mixed-format model files!