All filters compile for AVX-512 now.
However, I don't know if they work. We need a Xeon Skylake for
testing, because our code uses AVX-512 BW and DQ instructions, not
supported on Xeon Phi KNL. (Google Cloud Platform apparently received
Skylake servers in Feb; they're supposed to become widely available
later in the year.) Unit tests still pass, because our CPU dispatching
sees that the processor doesn't support BW/DQ instructions, even
though the compiler does, and the AVX impl is used at runtime (exactly
as we designed it, amazingly enough).

In principle, this is the final step in the vectorconf feature branch;
maybe a little cleanup left to do, and it'll be time to merge back to
`develop`.