SSE version of SSV, Viterbi filters use esl_sse_rightshift*() functions.
simdvec.md now documents that "left" and "right" refer to memory
order (low addresses to the left), and to the logic of our depiction
of striped DP algorithms. On little-endian architectures like x86,
this is the opposite of what a vector register looks like, so the
hardware instruction for what we call a "right" shift is a left shift.

Our left/right convention is chosen so that the algorithm description
is consistent and ISA-independent; it's up to the internals of an
individual ISA-specific implementation to implement the correct
hardware instructions.

I checked that using static inline esl_sse_rightshift*() functions
have no discernable impact on benchmark times, compared to not calling
functions. This lets us modularize and document the "right shift"
implementation in one place.