Fixes iss131: segfault if top seq has no reported domains
Sergey Ovchinnikov provides case where a sequence has no identified
domains, despite having region and envelope, because Decoding fails
with an eRANGE error as foreseen in 2008 {J3/119-121}. With ndom==0,
any code accessing a top hit structure with th->hit[h]->dcl[d] is
broken. Ovchinnikov example caused segfault in p7_tophits_Alignment()
-- see issue #131 -- but that turned out to be far from the only
problem.

Changed p7_tophits_Alignment() to be robust against the top hit having
no domains. (This did not fix the bug, but it's a good thing to do,
so I left it.)

Added a test for ndom > 0 in p7_Pipeline, where it was already testing
for nregions > 0 and nenvelopes > 0. Documented how ndom can be 0 even
for an identified envelope.