Try to ensure only a single instance of i19-hmmpgmd-ga.pl runs.
hmmpgmd_ga test can fail if more than one copy of i19-hmmpgmd-ga.pl is
run simultaneously, because only one instance of the hmmpgmd daemon
should be running. If you do 'make check' in parallel on multiple
builds to save time, it's possible (though rare) that the hmmpgmd_ga
test will be reached at the same time by two testsuite runs, causing a
difficult-to-reproduce failure.

To mitigate, added a flock of the i19-hmmpgmd-ga.pl script itself;
script tries multiple (10) times to get a flock on itself, waiting (5)
sec between attempts. This strategy works so long as it's the same
source copy of i19-hmmpgmd-ga.pl that's being run more than once,
which is the case for multiple 'make checks' in different build dirs
of the same source tree.