Fixes iss#140 'jackhmmer --fast segfaults' by removing --fast option.
jackhmmer is designed to assume that the profiles at every iteration
have identical lengths, built on the coordinate system of the original
query as the "consensus". It includes a query sequence trace in each
round's alignment, uses it as reference annotation, and uses "--hand"
model construction by default. If "--fast" is used, then profile
lengths change; when p7_tophits_Alignment() tries to combine the query
trace and top hits and they're on different coord systems, it'll
either corrupt the result or segfault.

Since the design of jackhmmer essentially requires --hand
construction, I removed the --fast and --symfrac options and their
documentation.

"Removing" them is fiddly. They're only present in jackhmmer in the
first place because of how p7_Builder() takes its configuration as a
commandline options structure, which was probably a bad idea. We have
to have --hand and --fast in the options structure that goes to
p7_Builder(). So the workaround to "remove" them is a) assign them to
help group 99 so they don't appear in the help documentation; b) add
code in jackhmmer's process_commandline() to issue an error if they're
changed from their defaults.

Additionally added a check in p7_tophits_Alignment() to catch this
kind of bug in the future.

The fix to iss#131 exposed this bug as a segfault; it was present in
earlier releases, but did not manifest itself such in an obvious way.