Fixes iss #142, compilation fails ungracefully on ppc64le
HMMER3's VMX implementation is bigendian, so it shouldn't be compiled
on ppc64le. Added configure.ac code to detect ppc64le, warn and exit.
Added documentation to say that ppc64 (BE) and powerpc (32 bit BE) are
supported but ppc64le is not, in install.tex.

Will leave issue #142 open nonetheless, because the opposite problem
remains in H4, which is the opposite: H4 assumes ppc64le only, and
fails on ppc64 or powerpc.