Improves flock()'ing in the i19-hmmpgmd-ga.pl test.
Commit 74b85d2 added a flock()'ing mechanism to ensure that only one
instance of the hmmpgmd daemon test runs at a time, where I flock'ed
the script itself. This worked at first, but after a CentOS 7 upgrade
on Odyssey, it stopped working on RC login head nodes. flock() is
known to be unreliable on some networked filesystems, and the way RC
mounts the eddyfs01 home dir has seemed fiddly, so I think that's the
problem.

Changed to flock'ing a sentinel file in /tmp, which is more likely to
be a local filesystem; this fixed the problem on Odyssey.