Begins work in "nwo" (new world order) directory, parallel to "src".
The idea is to pull H4 together now, at long last, and to do that in a
parallel directory (nwo), so I can build it cleanly from the bits and
bobs in src. The initial commit includes a framework of the new
`hmmer` single executable, which will have subcommands attached to it.

And what the hell, while I'm working on a clean-ish slate, I'll revert
to a Plan9 model. The lack of DI and ID transitions has caused
numerous difficulties downstream (in Infernal, in Elena's phylogenetic
work...), and may not be worth the trouble it's caused. We'll see if
this is a wise decision or not. Changing P7_HMM, the core-est of the
core data structures, may be crazy.

With the model itself changing, H4 functions will be prefaced h4_, not
p7_.

Yeah, I know, we'll see.