Fixes a set of qsort() bugs, related to fixes in H3, Easel, Infernal.
A qsort() helper function needs to return 0 when its keys are
equal. We had multiple places where qsort helpers are only returning
-1/1. First reported in EddyRivasLab/infernal#11 by Yuri Gribov
(yugr), author of SortCheck. In May 2018, EPN committed Infernal
fixes, and sent a pull request (#139) with H3 fixes. Belatedly, here
are the analogous H4 fixes.

Also, adds -L flag to the find's in makeTAGS.sh, so the TAGS file
includes lib/easel when that's a symlink.