hmmstat output documented in userguide.
Adds a section of the userguide showing an example of hmmstat, and
documenting its output. (The output fields were documented in the man
page, just not in the userguide.) Requested by Sarah Berkemer (Leipzig).

Adds a --nodelete option to gen-inclusions.py, so you can have it leave
its tmpfiles and symlinks around; convenient when developing new inclusions
(like now, for hmmstat).

Adds a Makefile for the userguide/inclusions directory, mostly to make
it easier to remember that all that happens in this directory is an
occasional `./gen-inclusions.py`, usually at release time.