Changes userguide documentation of STATS lines in a profile.
User guide called the STATS lines "optional". This is literally true
in the sense that the profile can be parsed without them, but if the
E-value parameters aren't present, then hmmsearch/hmmscan will
silently give no hits, because the acceleration filters need to be
able to calculate P-values.

Updated the documentation to call them "mandatory", with a side note
explaining why.

More robustly, hmmsearch and hmmscan should check that E-value
parameters are available in a query model - leave that for H4.