Improves flock technique ensuring only one test daemon runs at a time.
i19 and i22 itests aim to ensure that only one daemon is running at a
time on a given machine, across all simultaneous `make check`
invocations, across any number of different HMMER build trees *and
users*. To do this, we create a globally shared lockfile
"/tmp/esl-hmmpgmd-test.lock", and have the daemon test scripts grab an
exclusive flock on this file. The lockfile must be writable by the
itest, because exclusive flocks only work on filehandles open for
writing. The lockfile must persist while any itests could be running,
so we don't have the itests clean it up.

I hadn't previously considered what happens across different users.
Especially if a user with a restrictive umask (e.g. Tom) creates the
lockfile first with a 'make check', subsequent users on the same
machine (e.g. eddylogin ody head node) won't be able to open the file.

To fix this, added `umask 0011` to override user umask, and an
explicit `chmod 0666` on the lockfile too for good measure.