Revise hmmsearcht workflow for speed, reduce ESL_SQ allocations.
Previous implementation read in an entire sequence (up to a chromosome)
and translated the whole thing en masse, which required a massive
number of ESL_SQ mallocs. This commit fixes this in the serial (single
thread) workflow, reading sequence windows as in nhmmer, and reusing
ESL_SQ objects between windows.

Known problems:
(1) segfault due to an as-yet unidentified memory overrun. I am committing
here to move to a better system for testing
(2) multi-threaded version not yet updated
(3) summary report lists an excess of Target sequences