Merging the new serialization scheme into develop.
This change replaces the ad-hoc serialization scheme used in hmmpgmd with a
more regimented one, in which each class that needs to be sent via sockets
has foo_Serialize() and foo_Deserialize() functions that can be used to build
buffers of data that can be sent via sockets.  These functions comply with
internet standard (big-endian) ordering of multi-byte objects.