Fixes iss #159 : nhmmer reports overlapping envelopes on reverse strand.
p7_tophits_RemoveDuplicates() depends on a particular sort order for
hit coords, and the logic in `hit_sorter_by_seqidx_aliposition()` (for
nhmmer) and `hit_sorter_by_modelname_aliposition()` (for nhmmscan)
failed to consider what happens with hits on the opposite (crick)
strand. Fixed, and improved documentation around the issue.

[xref SRE:2019/0531-nhmmer-iss159; includes detailed example.]